+++ src/components/KPDatePicker/index.scss
@@ -431,7 +431,7 @@
             content: "";
             width: 16px;
             height: 16px;
-            background-image: url('../img/ArrowDownRed.svg');
+            background-image: url('../../assets/img/ArrowDownRed.svg');
             position: absolute;
             right: 0;
             top: 2px;
