+++ src/containers/DatePicker/index.js
@@ -1,0 +1,3 @@
+import DatePicker from './DatePicker';
+
+export default DatePicker;
